Move Foundation namespace to root Hyde namespace
