+++ packages/framework/src/Foundation/Concerns/BaseFoundationCollection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Foundation\Concerns;
+namespace Hyde\Foundation\Concerns;
 
 use Hyde\Framework\HydeKernel;
 use Illuminate\Support\Collection;
@@ -10,9 +10,9 @@ use Illuminate\Support\Collection;
 /**
  * @internal Base class for the kernel auto-discovery collections.
  *
- * @see \Hyde\Framework\Foundation\FileCollection
- * @see \Hyde\Framework\Foundation\PageCollection
- * @see \Hyde\Framework\Foundation\RouteCollection
+ * @see \Hyde\Foundation\FileCollection
+ * @see \Hyde\Foundation\PageCollection
+ * @see \Hyde\Foundation\RouteCollection
  */
 abstract class BaseFoundationCollection extends Collection
 {
