+++ packages/framework/src/Foundation/Concerns/ForwardsFilesystem.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Foundation\Concerns;
+namespace Hyde\Foundation\Concerns;
 
 /**
  * @internal Single-use trait for the HydeKernel class.
