+++ packages/framework/src/Foundation/Filesystem.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Foundation;
+namespace Hyde\Foundation;
 
 use Hyde\Framework\HydeKernel;
 use Hyde\Framework\Models\Pages\BladePage;
