+++ packages/framework/src/Foundation/RouteCollection.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Hyde\Framework\Foundation;
+namespace Hyde\Foundation;
 
+use Hyde\Foundation\Concerns\BaseFoundationCollection;
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Framework\Foundation\Concerns\BaseFoundationCollection;
 use Hyde\Framework\Models\Support\Route;
 
 /**
  * Pseudo-Router for Hyde.
  *
- * @see \Hyde\Framework\Foundation\PageCollection
+ * @see \Hyde\Foundation\PageCollection
  * @see \Hyde\Framework\Testing\Feature\RouteTest
  * @see \Hyde\Framework\Testing\Feature\RouteCollectionTest
  *
