+++ packages/framework/src/Hyde.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde;
 
+use Hyde\Foundation\FileCollection;
+use Hyde\Foundation\PageCollection;
+use Hyde\Foundation\RouteCollection;
 use Hyde\Framework\Concerns\HydePage;
-use Hyde\Framework\Foundation\FileCollection;
-use Hyde\Framework\Foundation\PageCollection;
-use Hyde\Framework\Foundation\RouteCollection;
 use Hyde\Framework\Helpers\Features;
 use Hyde\Framework\HydeKernel;
 use Hyde\Framework\Models\Support\Route;
