+++ packages/framework/tests/Feature/Foundation/FilesystemTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature\Foundation;
 
-use Hyde\Framework\Foundation\Filesystem;
+use Hyde\Foundation\Filesystem;
 use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
@@ -14,7 +14,7 @@ use Hyde\Testing\TestCase;
 
 /**
  * @covers \Hyde\Framework\HydeKernel
- * @covers \Hyde\Framework\Foundation\Filesystem
+ * @covers \Hyde\Foundation\Filesystem
  */
 class FilesystemTest extends TestCase
 {
@@ -91,7 +91,7 @@ class FilesystemTest extends TestCase
 
     public function test_vendor_path_method_exists()
     {
-        $this->assertTrue(method_exists(Filesystem::class, 'vendorPath'));
+        $this->assertTrue(method_exists(\Hyde\Foundation\Filesystem::class, 'vendorPath'));
     }
 
     public function test_vendor_path_method_returns_string()
@@ -113,7 +113,7 @@ class FilesystemTest extends TestCase
     public function test_copy_method()
     {
         touch(Hyde::path('foo'));
-        $this->assertTrue(method_exists(Filesystem::class, 'copy'));
+        $this->assertTrue(method_exists(\Hyde\Foundation\Filesystem::class, 'copy'));
         $this->assertTrue(Hyde::copy('foo', 'bar'));
         $this->assertFileExists(Hyde::path('bar'));
         unlink(Hyde::path('foo'));
