+++ packages/framework/tests/Feature/Foundation/HyperlinksTest.php
@@ -4,22 +4,21 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature\Foundation;
 
-use Hyde\Framework\Foundation\Hyperlinks;
 use Hyde\Framework\HydeKernel;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Framework\Foundation\Hyperlinks
+ * @covers \Hyde\Foundation\Hyperlinks
  */
 class HyperlinksTest extends TestCase
 {
-    protected Hyperlinks $class;
+    protected \Hyde\Foundation\Hyperlinks $class;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->class = new Hyperlinks(HydeKernel::getInstance());
+        $this->class = new \Hyde\Foundation\Hyperlinks(HydeKernel::getInstance());
     }
 
     public function test_image_helper_gets_relative_web_link_to_image_stored_in_site_media_folder()
